LayoutTests/sputnik/Conformance/07_Lexical_Conventions/7.9_Automatic_Semicolon_Insertion/S7.9_A4-expected.txt

CONSOLE MESSAGE: line 79: SyntaxError: Unexpected number '1'
S7.9_A4

PASS Expected parsing failure

TEST COMPLETE
